requirements.txt

openai>=1.2.0
streamlit
pillow
